refactor(middleware): extract the policy/step-up gate to a sub-factory
Move withPolicyGate into with-kya-os.policy-gate.ts behind createPolicyGate(deps, { attachOutcomeProof }); add the shared AttachOutcomeProof contract type to with-kya-os.deps.ts and prune the now-dead policy imports from the orchestrator. No behaviour change; 1607 tests green. Phase 2b.
